refactor: delegate functions to services and improve response handling
