--- spring/src/main/java/com/iuha/api/service/ChatRoomServiceImpl.java
@@ -1,4 +1,0 @@
-package com.iuha.api.service;
-
-public class ChatRoomServiceImpl implements ChatRoomService{
-}
